% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_transitory_shock.R
\name{add_transitory_shock}
\alias{add_transitory_shock}
\title{add transitory shock to the data}
\usage{
add_transitory_shock(data, shock_date, restricted = FALSE)
}
\arguments{
\item{data}{is the data frame that is to be used for modeling}

\item{shock_date}{is a date object with the date for the transitory shock}

\item{restricted}{is a boolean indicating whether the}
}
\value{
the input xts object with the added
}

