% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endogenous.R
\name{get_endogenous}
\alias{get_endogenous}
\title{Get endogenous variables from data set}
\usage{
get_endogenous(data)
}
\arguments{
\item{data}{is a data set with the name codes of
the Rcivecm package}
}
\value{
an xts object with the endogenous variable parts
        of the original data
}

