% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exogenous.R
\name{get_exogenous}
\alias{get_exogenous}
\title{Extract exogenous variables from an xts data object}
\usage{
get_exogenous(data)
}
\arguments{
\item{data}{is an xts object with variable names following
the scheme in Rcivecm}
}
\value{
an xts object with the exogenous variable parts of the original data.
}

