% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_restricted.R
\name{get_restricted}
\alias{get_restricted}
\title{Get exogenous variables restricted to the cointegration vector}
\usage{
get_restricted(data)
}
\arguments{
\item{data}{is an xts object with the naming conventions of
Rcivecm.}
}
\value{
an xts object with the exogenous variables that
        are restricted to the cointegration space
}

