% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Z2.R
\name{get_Z2}
\alias{get_Z2}
\title{Get Z2}
\usage{
get_Z2(endogenous, exogenous, lags)
}
\arguments{
\item{endogenous}{is the endogenous variables part of the data}

\item{lags}{is the number of lags in levels}

\item{exogensous}{is the exogenous variables part of the data}
}
\value{
Z2
}

