% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interest_rates}
\alias{interest_rates}
\title{A dataset with daily observations of US interest rates
       from 04-Jan-2010 to 31-12-2014.}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 1251 rows and 3 columns.}
\usage{
interest_rates
}
\description{
The data is downloaded from Quandl using the codes
             \code{'YC/USA10Y','YC/USA5Y','YC/USA2Y'}. The data
             is daily spanning from 2010 to 2016.
}
\keyword{datasets}

